[front] chore: Migrate Salesforce MCP
