+++ front/lib/api/actions/servers/salesforce/helpers.ts
@@ -1,0 +1,59 @@
+import type { Connection } from "jsforce";
+import jsforce from "jsforce";
+
+import { MCPError } from "@app/lib/actions/mcp_errors";
+import type {
+  ToolHandlerExtra,
+  ToolHandlerResult,
+} from "@app/lib/actions/mcp_internal_actions/tool_definition";
+import logger from "@app/logger/logger";
+import { Err, normalizeError } from "@app/types";
+
+export const SF_API_VERSION = "57.0";
+
+export async function withAuth(
+  extra: ToolHandlerExtra,
+  action: (conn: Connection) => Promise<ToolHandlerResult>
+): Promise<ToolHandlerResult> {
+  const accessToken = extra.authInfo?.token;
+  const instanceUrl = extra.authInfo?.extra?.instance_url;
+
+  if (typeof instanceUrl !== "string") {
+    return new Err(new MCPError("Missing or invalid instance_url in authInfo"));
+  }
+
+  const conn = new jsforce.Connection({
+    instanceUrl,
+    accessToken,
+    version: SF_API_VERSION,
+  });
+
+  try {
+    await conn.identity();
+  } catch (error) {
+    return new Err(
+      new MCPError(
+        `Failed to authenticate with Salesforce: ${normalizeError(error).message}`
+      )
+    );
+  }
+
+  return action(conn);
+}
+
+export function logAndReturnError({
+  error,
+  params,
+  message,
+}: {
+  error: unknown;
+  params: Record<string, unknown>;
+  message: string;
+}): ToolHandlerResult {
+  const normalizedError = normalizeError(error);
+  logger.error(
+    { params, error: normalizedError.message },
+    `Salesforce MCP: ${message}`
+  );
+  return new Err(new MCPError(`${message}: ${normalizedError.message}`));
+}
